requirements_prod.txt
streamlit==1.28.1
feedparser==6.0.10
requests==2.31.0
pandas==2.1.0
numpy==1.24.0
python-dateutil==2.8.2
plotly==5.17.0
